Fixed server issue by switching to NimBLE library rather than Arduino's default BLE. This uses much less RAM and flash space which was causing the issue. Also optimised and fixed the delete UI element code for the config server, it is now tested and working.

+++ src/utils.cpp
@@ -32,7 +32,7 @@ const char* Utils::getUiScreenFileName(UiScreen screen) {
 }
 
 const char* Utils::getUiScreenFileName(uint8_t screen) {
-  return (kUiScreenFilePrefix + String(static_cast<uint8_t>(screen)) + "." + kUiScreenFileType).c_str();
+  return (kUiScreenFilePrefix + String(screen) + "." + kUiScreenFileType).c_str();
 }
 
 }
